Version 1.2.2-and-a-bit

+++ File.h
@@ -11,8 +11,8 @@
 	NSString *filePath;
 	NSString *displayName;
 	
-	unsigned long byteSize;
-    unsigned long byteSizeOptimized;	
+	NSUInteger byteSize;
+    NSUInteger byteSizeOptimized;	
     NSString *bestToolName;
 	double percentDone;
 	
@@ -23,9 +23,9 @@
     
     NSMutableArray *workers;
     
-	unsigned int workersActive;
-	unsigned int workersFinished;
-	unsigned int workersTotal;
+	NSUInteger workersActive;
+	NSUInteger workersFinished;
+	NSUInteger workersTotal;
     
     NSOperationQueue *fileIOQueue;
 }
@@ -34,19 +34,19 @@
 
 -(void)enqueueWorkersInCPUQueue:(NSOperationQueue *)queue fileIOQueue:(NSOperationQueue *)fileIOQueue;
 
--(void)setFilePathOptimized:(NSString *)f size:(unsigned long)s toolName:(NSString*)s;
+-(void)setFilePathOptimized:(NSString *)f size:(NSUInteger)s toolName:(NSString*)s;
 
 -(id)initWithFilePath:(NSString *)name;
 -(id)copyWithZone:(NSZone *)zone;
--(void)setByteSize:(unsigned long)size;
--(void)setByteSizeOptimized:(unsigned long)size;
+-(void)setByteSize:(NSUInteger)size;
+-(void)setByteSizeOptimized:(NSUInteger)size;
 -(BOOL)isOptimized;
 
 -(void)setFilePath:(NSString *)s;
 
 -(NSString *)fileName;
 
-@property (assign) unsigned long byteSize, byteSizeOptimized;
+@property (assign) NSUInteger byteSize, byteSizeOptimized;
 @property (retain) NSString *statusText, *filePath, *displayName;
 @property (retain) NSImage *statusImage;
