+++ Workers/PngCrushWorker.m
@@ -65,7 +65,7 @@
 	
 	if (![task terminationStatus])
 	{
-		unsigned long fileSizeOptimized;
+		NSUInteger fileSizeOptimized;
 		if (fileSizeOptimized = [File fileByteSize:temp])
 		{
 			[file setFilePathOptimized:temp	size:fileSizeOptimized toolName:[self className]];
